rename phi4o with phi4mm
